v0.100
The first commit of the label maker project for Bedloft.com.

Purpose of this project is to allow the user to upload an excel
file, the web application will read then write a new file which
will either be labels or a roster.

Current version only includes the first step towards the creation
of the labels.

-AM