v0.200
The core of this version is an updated LabelController that optimizes
the PhpSpreadsheet reader with an excel "template" that ensures
the same look for the label. Currently it only takes the first
name and last name and the logo at the bottom is not 100%, but
future updates will resolve this. Focus is still on the labels.

No major updates to the "upload" view. Soon the user will be able
enter data and select what they want generated (label or roster).

Purpose of this project is to allow the user to upload an excel
file, the web application will read then write a new file which
will either be labels or a roster.

Current version only includes the first step towards the creation
of the labels.

-AM